requirements.txt
google-cloud-bigquery>=3.11.0
pandas>=1.5.0
pyarrow>=10.0.0
google-auth>=2.16.0
google-cloud-storage>=2.7.0
